Try moving `textStyle` to Element
